+++ Tests/RoktUXHelperTests/UI/Components/TestBasicTextComponent.swift
@@ -74,7 +74,10 @@ final class TestBasicTextComponent: XCTestCase {
             .frame(width: 350)
         
         let hostingController = UIHostingController(rootView: view)
-        assertSnapshot(of: hostingController, as: .image(on: snapshotDevice))
+        assertSnapshot(
+            of: hostingController,
+            as: .image(on: snapshotDevice, precision: snapshotPrecision, perceptualPrecision: snapshotPerceptualPrecision)
+        )
     }
     
     func get_model() throws -> BasicTextViewModel {