+++ Tests/RoktUXHelperTests/UI/Components/TestZStackComponent.swift
@@ -53,7 +53,9 @@ final class TestZStackComponent: XCTestCase {
             .frame(width: 350, height: 350)
 
         let hostingController = UIHostingController(rootView: view)
-        assertSnapshot(of: hostingController, as: .image(on: snapshotDevice))
+        assertSnapshot(of: hostingController,
+                       as: .image(on: snapshotDevice, precision: snapshotPrecision,
+                                  perceptualPrecision: snapshotPerceptualPrecision))
     }
 
     // MARK: - Helpers